// apps/frontend/src/types/bindings/Auth.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type AuthResponse = { access_token: string, refresh_token: string, token_type: string, expires_in: number, };

export type LoginRequest = { email: string, password: string, };

export type RefreshRequest = { refresh_token: string, };

export type RegisterRequest = { email: string, password: string, confirmPassword: string, };

export type UserResponse = { id: string, email: string, first_name: string, last_name: string, };